+++ lib/src/util.dart
@@ -14,6 +14,9 @@
 
 import 'dart:async';
 
+import 'package:charcode/charcode.dart';
+import 'package:string_scanner/string_scanner.dart';
+
 /// Runs [onFinally] after [callback] completes, whether it returns a [Future]
 /// or a synchronous value.
 T tryFinally<T>(T callback(), void onFinally()) {
@@ -28,6 +31,39 @@ T tryFinally<T>(T callback(), void onFinally()) {
   if (result is Future) return result.whenComplete(onFinally) as T;
   onFinally();
   return result;
+}
+
+/// Replaces [match] with a [replacement] string in the style of
+/// [LineStreamExtensions.replace].
+String replaceMatch(Match match, String replacement) {
+  var scanner = SpanScanner(replacement);
+  var buffer = StringBuffer();
+
+  while (!scanner.isDone) {
+    var next = scanner.readChar();
+    if (next != $backslash) {
+      buffer.writeCharCode(next);
+      continue;
+    }
+
+    if (scanner.isDone) break;
+
+    next = scanner.readChar();
+    if (next >= $0 && next <= $9) {
+      var groupNumber = next - $0;
+      if (groupNumber > match.groupCount) {
+        scanner.error("RegExp doesn't have group $groupNumber.",
+            position: scanner.position - 2, length: 2);
+      }
+
+      var group = match[groupNumber];
+      if (group != null) buffer.write(group);
+    } else {
+      buffer.writeCharCode(next);
+    }
+  }
+
+  return buffer.toString();
 }
 
 /// Stream extensions used internally within `cli_script`, not for public